Bugfix endless url loop for faulty breakpoint
There was a small bug where when the first url did not find any result,
instead of continueing to the second one it did a break and went out of
the loop and starting again at the first url, those creating endles
loop. Therefore had to replace the break with continue.